删除war包

+++ vnc/vnc.sh
@@ -1,7 +1,22 @@
 #! /bin/sh
-sudo cp ./vnc.pass /etc/
-sudo cp ./x11vnc.service /usr/lib/systemd/system/
-sudo systemctl daemon-reload
-sudo systemctl start x11vnc
+
+
+if [ ! -f /etc/vnc.pass ];then
+    sudo cp ./vnc.pass /etc/
+fi
+
+if [ ! -f /usr/lib/systemd/system/x11vnc.service ]:then
+    sudo cp ./x11vnc.service /usr/lib/systemd/system/
+    sudo systemctl daemon-reload
+    sudo systemctl start x11vnc
+    sudo systemctl enable x11vnc
+else
+    sudo systemctl status x11vnc
+fi
+
+
+
+
+
